Move Threadpool to separate file
